+++ src/certsling/tests/test_generate.py
@@ -10,7 +10,10 @@ def generate(base, ca):
 
 def test_user_gen(base, generate, verify_crt_true, yesno_true):
     assert list(base.iterdir()) == []
-    generate(domains=['example.com'], regenerate=False, update_registration=False)
+    domains = ['example.com']
+    generate(
+        main=domains[0], domains=domains,
+        regenerate=False, update_registration=False)
     fns = list(x.name for x in base.iterdir())
     assert 'user.key' in fns
     assert 'user.pub' in fns